Implement remove_bytes_from_file

+++ eof_tools.h
@@ -1,7 +1,8 @@
 #include <string>
 #include <fstream>
+#include <sys/stat.h>
 using namespace std;
 
 bool file_ends_in(string filename, string ending);
 void append_to_file(string filename, string ending);
-void remove_bytes_from_file(string filename, int bytes);
+void remove_bytes_from_file(string filename, uint64_t bytes);
